% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pubmap.R
\name{abstractWordcloud}
\alias{abstractWordcloud}
\title{abstractWordcloud}
\usage{
abstractWordcloud(abstract, n = 35)
}
\arguments{
\item{abstract}{abstract text}
}
\value{
plot
}
\description{
plot word cloud of abstract
}
\author{
ygc
}

