% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{barplot}
\usage{
barplot(data, title = "", xlab = "", ylab = "", barIndex = NULL,
  label = "*")
}
\arguments{
\item{data}{data.frame}

\item{title}{plot title}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{barIndex}{index of bar to add annotation of statistical significance}

\item{label}{statistical significant label, eg. *, **}
}
\value{
ggplot2 object
}
\description{
barplot with errorbar
}
\details{
barplot
}
\examples{
df <- data.frame(Normal=c(0.83,0.79,0.99,0.69), Cancer=c(0.56,0.56,0.64,0.52))
barplot(df, barIndex=c(1,2), label="*")
}
\author{
Guangchuang Yu
}

