% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pubmap.R
\name{getPubmed}
\alias{getPubmed}
\title{getPubmed}
\usage{
getPubmed(searchTerm)
}
\arguments{
\item{searchTerm}{search term}
}
\value{
data.frame
}
\description{
query pubmed information
}
\author{
ygc
}

