% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPubmedTrend}
\alias{plotPubmedTrend}
\title{plotPubmedTrend}
\usage{
plotPubmedTrend(x)
}
\arguments{
\item{x}{data.frame or a name list of data.frame, which output by getPubmedTrend}
}
\value{
figure
}
\description{
visualize pubmed trend
}
\author{
ygc
}

