% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pubmap.R
\name{getPubmedTrend}
\alias{getPubmedTrend}
\title{getPubmedTrend}
\usage{
getPubmedTrend(searchTerm, year)
}
\arguments{
\item{searchTerm}{search term}

\item{year}{year vector}
}
\value{
plot
}
\description{
query pubmed record numbers of search term
}
\author{
ygc
}

